Rename collection to partition (#1)
* rename collection to partition

* rename collection to partition

* rename partition to table

* fix: added TpPartition hive field

---------
