--- azure_collection/activity_log_collection_config.go
@@ -1,7 +1,0 @@
-package azure_collection
-
-type ActivityLogCollectionConfig struct{}
-
-func (a *ActivityLogCollectionConfig) Validate() error {
-	return nil
-}
