+++ azure_table/activity_log_table_config.go
@@ -1,0 +1,7 @@
+package azure_table
+
+type ActivityLogTableConfig struct{}
+
+func (a *ActivityLogTableConfig) Validate() error {
+	return nil
+}
